# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Forafix API Documentation",
        "description": "Marketplace API with Wallet, Escrow and Booking management",
        "contact": {
            "email": "support@forafix.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/upload/avatar": {
            "post": {
                "tags": [
                    "Media"
                ],
                "summary": "Upload and update user profile picture",
                "operationId": "9d9f7dd6bad72fa6048bfa400a2ec74a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/password/update": {
            "post": {
                "tags": [
                    "Security"
                ],
                "summary": "Change account password",
                "operationId": "f644d024a05857e0ed3e85a6507fb398",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/payments/pay-with-wallet": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Pay for a booking using wallet balance (Escrow)",
                "operationId": "478e6d7aa87755bfe31f0f30d91a6a78",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "booking_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment successful"
                    },
                    "400": {
                        "description": "Invalid state or insufficient balance"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/services/{slug}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get specific service by slug",
                "operationId": "14262bc076f0332906ac8f923d209c09",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}